Fix store map being reset on DeleteAll calls
The internal store map was unconditionally reset on all calls to
DeleteAll.

In a circumstance where the user has multiple stores open from different
types - such as one ReadOnly and one RawData store - calling DeleteAll
with the store type of "ReadOnly" would then prevent the user from
being able to call the CloseStore method with the handle to the RawData
backed store.