Various fixes for the Delete store method
The Delete store method had a few issues which prevented its initial
implementation from actually working.

Firstly, the RegisterStoreType function wasn't copying the Delete method
from the supplied prototype to the table it creates.

Next, the Delete verb was incorrectly indexing the prototypes table by
the ID of the given store, instead of the store type.

Finally, with those two issues corrected there was a typo in the call
to DeArchive which used a lowercase "A" and the parameter supplied to it
was the saved variable table for the store, not the data field.